backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlmodel>=0.0.16
asyncpg>=0.29.0
python-multipart>=0.0.9
pandas>=2.0.0
reportlab>=4.0.0
# psycopg2-binary>=2.9.0
# psycopg[binary]>=3.1
# psycopg>=3.1
psycopg-binary>=3.1.8
python-dotenv>=1.0.0
alembic>=1.13.0 